refactor(core): split cache agent modules
