+++ crates/mbx-cache-core/src/agent.rs
@@ -8,7 +8,6 @@ use crate::{
 use eyre::{Context, Result, bail};
 use futures_util::{FutureExt, StreamExt, future::BoxFuture, stream};
 use log::warn;
-use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
 use std::fs;
 use std::path::{Path, PathBuf};
@@ -16,6 +15,27 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, Mutex, Weak};
 use std::time::{Duration, Instant, SystemTime};
 use tokio::io::{AsyncBufRead, AsyncBufReadExt, AsyncRead, AsyncWrite, AsyncWriteExt, BufReader};
+
+mod file_digest;
+mod manifest;
+mod prefetch;
+mod stats;
+mod wire;
+
+pub use file_digest::{
+    FileDigestCache, FileDigestScope, FileIdentity, NoFileDigestCache, RecordedFileDigest,
+};
+pub use manifest::{is_task_identity, task_manifest_actions};
+use manifest::{
+    merge_remote_task_manifest, merge_task_manifests, task_manifest_dir, validate_task_identity,
+    validate_task_manifest,
+};
+pub use stats::{AgentStats, CompilerStats};
+use wire::MAX_REQUEST_BYTES;
+pub use wire::{
+    AGENT_PROTOCOL_VERSION, AgentEvent, AgentEventObserver, AgentRequest, AgentResponse,
+    RestoreStats,
+};
 
 const MAX_EXECUTABLE_IDENTITIES: usize = 64;
 const MAX_EXECUTABLE_IDENTITY_SIZE: usize = 64 * 1024;
@@ -56,467 +76,6 @@ pub struct AgentRemoteCache {
     pub mode: RemoteCacheMode,
     /// Directory used for verified downloads before CAS ingestion.
     pub staging_dir: PathBuf,
-}
-
-/// Wire protocol version used between an in-process cache agent and its shims.
-pub const AGENT_PROTOCOL_VERSION: u8 = 5;
-/// Largest single protocol request the agent will read.
-///
-/// Requests are small JSON objects; the largest legitimate ones carry an output
-/// tree or a batch of digests, which stay far below this.
-const MAX_REQUEST_BYTES: usize = 16 * 1024 * 1024;
-
-/// A request accepted by the task-scoped cache agent.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum AgentRequest {
-    /// Negotiate protocol and application versions.
-    Hello {
-        /// Agent protocol version understood by the caller.
-        protocol: u8,
-        /// Human-readable mbx client version.
-        client_version: String,
-    },
-    /// Begin a prediction-manifest run for a stable Cargo or task identity.
-    BeginTask {
-        /// Stable 64-character lowercase hexadecimal task identity.
-        task: String,
-    },
-    /// Commit predictions collected by an earlier [`Self::BeginTask`].
-    CommitTask {
-        /// Opaque run identifier returned by the agent.
-        run: String,
-    },
-    /// Resolve a blob to a session-verified local CAS path.
-    FindBlob {
-        /// Blob to resolve.
-        digest: CacheDigest,
-    },
-    /// Resolve blobs to session-verified local CAS paths.
-    FindBlobs {
-        /// Blobs to resolve, preserving request order in the response.
-        digests: Vec<CacheDigest>,
-    },
-    /// Import a file into the local content-addressed store.
-    StoreBlob {
-        /// Digest the source must match.
-        digest: CacheDigest,
-        /// File to verify and import.
-        source: PathBuf,
-    },
-    /// Look up an action-result record.
-    FindActionResult {
-        /// Action digest to resolve.
-        action: CacheDigest,
-    },
-    /// Account for a successfully restored cache hit.
-    RecordActionHit {
-        /// Action that supplied the outputs.
-        action: CacheDigest,
-        /// Restoration work performed by the adapter.
-        restore: RestoreStats,
-        /// Compiler crate name, when the invocation supplied one.
-        #[serde(default)]
-        crate_name: Option<String>,
-    },
-    /// A compilation the adapter declined to cache, grouped by reason.
-    RecordBypass {
-        /// Stable, low-cardinality bypass-reason name.
-        kind: String,
-    },
-    /// A compilation the adapter could not look up, having no key to look up
-    /// with. Distinct from a bypass: these are cached once compiled.
-    RecordUnconsulted,
-    /// Account for one real compiler invocation performed by the adapter.
-    RecordCompilerInvocation {
-        /// Stable outcome category such as `miss`, `unconsulted`, `bypass`, or
-        /// `incremental` for a compilation the adapter deliberately ran with
-        /// incremental state instead of publishing.
-        outcome: String,
-        /// Compiler crate name, when the invocation supplied one.
-        crate_name: Option<String>,
-        /// Wall time spent running the compiler.
-        duration_ns: u64,
-    },
-    /// Account for a cache hit that was rebuilt for correctness verification.
-    RecordActionVerification {
-        /// Whether rebuilt and cached outputs matched.
-        matched: bool,
-        /// Restoration work performed before rebuilding.
-        restore: RestoreStats,
-    },
-    /// Store an action-result record locally and enqueue remote publication.
-    StoreActionResult {
-        /// Action-result record to store.
-        result: RemoteActionResult,
-    },
-    /// Find an earlier input prediction for a task and invocation.
-    FindActionPrediction {
-        /// Stable task identity.
-        task: String,
-        /// Digest of the compiler invocation without discovered inputs.
-        invocation: CacheDigest,
-    },
-    /// Record an input prediction after a successful compilation.
-    RecordActionPrediction {
-        /// Stable task identity.
-        task: String,
-        /// Adapter-owned prediction record.
-        prediction: ActionPrediction,
-    },
-    /// Find cached identity output for an executable and environment.
-    FindExecutableIdentity {
-        /// Executable whose identity command would run.
-        executable: PathBuf,
-        /// Environment variables affecting identity output.
-        environment: BTreeMap<String, Option<String>>,
-    },
-    /// Cache identity output for an executable and environment.
-    StoreExecutableIdentity {
-        /// Executable whose identity command ran.
-        executable: PathBuf,
-        /// Environment variables affecting identity output.
-        environment: BTreeMap<String, Option<String>>,
-        /// Captured identity-command standard output.
-        stdout: Vec<u8>,
-    },
-    /// Surface a shim diagnostic through the session that owns the build.
-    ///
-    /// A shim must not print diagnostics itself: its stderr belongs to the
-    /// compiler it stands in for, and build scripts read that stream as part
-    /// of the compiler's answer -- cc-rs treats any stderr output from a flag
-    /// probe as "unsupported", which changes the flags of every compilation
-    /// that follows and, with them, every action key the build produces.
-    ///
-    /// Appended rather than grouped with the other `Record*` requests it
-    /// belongs beside: these variants carry no `repr`, so inserting one moves
-    /// the discriminant of every variant after it, and a break nobody asked
-    /// for is worth less than the grouping. New variants go here.
-    RecordWarning {
-        /// Human-readable single-line diagnostic.
-        message: String,
-    },
-    /// Find session-recorded digests for files with these identities.
-    FindFileDigests {
-        /// What the recorded digests may stand in for.
-        scope: FileDigestScope,
-        /// File identities to resolve, preserving request order in the
-        /// response.
-        files: Vec<FileIdentity>,
-    },
-    /// Record digests of files a shim hashed or wrote this session.
-    RecordFileDigests {
-        /// What the recorded digests may stand in for.
-        scope: FileDigestScope,
-        /// Hashed files and the identities their digests describe.
-        entries: Vec<RecordedFileDigest>,
-    },
-}
-
-/// Local output restoration work performed by one action-cache adapter hit.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct RestoreStats {
-    /// Cumulative time spent materializing and validating output files.
-    pub duration_ns: u64,
-    /// Compiler wall time recorded when this action was originally produced.
-    /// Zero means no timing hint was available.
-    pub avoided_compiler_duration_ns: u64,
-    /// Number of compiler output files restored.
-    pub output_files: u64,
-    /// Declared size of compiler output files restored.
-    pub output_bytes: u64,
-    /// Number of restored output files that share data blocks with the CAS.
-    pub reflinked_output_files: u64,
-    /// Declared size of restored outputs that share data blocks with the CAS.
-    pub reflinked_output_bytes: u64,
-    /// Number of restored output files that required a byte-for-byte copy.
-    pub copied_output_files: u64,
-    /// Declared size of restored outputs that required a byte-for-byte copy.
-    pub copied_output_bytes: u64,
-    /// Number of output files already in place with the cached contents, kept
-    /// rather than rewritten.
-    pub reused_output_files: u64,
-    /// Declared size of outputs kept in place rather than rewritten.
-    pub reused_output_bytes: u64,
-}
-
-/// One accounted cache decision, as it happens.
-///
-/// The agent already folds every one of these into [`AgentStats`]; an observer
-/// sees the same decisions individually, before that summing loses the crate
-/// they belong to. Delivered synchronously from the request handler, so an
-/// observer that blocks slows the build it is watching.
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum AgentEvent {
-    /// An action's outputs were restored from cache.
-    ActionHit {
-        /// Compiler crate name, when the invocation supplied one.
-        crate_name: Option<String>,
-        /// Restoration work performed by the adapter.
-        restore: RestoreStats,
-    },
-    /// A compilation the adapter declined to cache.
-    Bypass {
-        /// Stable, low-cardinality bypass-reason name.
-        kind: String,
-    },
-    /// A compilation no lookup was possible for.
-    Unconsulted,
-    /// A real compiler invocation ran.
-    CompilerInvocation {
-        /// Stable outcome category such as `miss`, `unconsulted`, or `bypass`.
-        outcome: String,
-        /// Compiler crate name, when the invocation supplied one.
-        crate_name: Option<String>,
-        /// Wall time spent running the compiler.
-        duration_ns: u64,
-    },
-    /// A hit was rebuilt to verify it.
-    Verification {
-        /// Whether rebuilt and cached outputs matched.
-        matched: bool,
-        /// Restoration work performed before rebuilding.
-        restore: RestoreStats,
-    },
-    /// A shim reported a diagnostic for the session to surface.
-    Warning {
-        /// Human-readable single-line diagnostic.
-        message: String,
-    },
-}
-
-/// A sink for [`AgentEvent`]s observed during one session.
-pub trait AgentEventObserver: Send + Sync {
-    /// Handle one event. Must not panic, and should not block.
-    fn event(&self, event: AgentEvent);
-}
-
-/// A response returned by the task-scoped cache agent.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum AgentResponse {
-    /// Successful protocol negotiation.
-    Hello {
-        /// Agent protocol version.
-        protocol: u8,
-        /// Human-readable agent version.
-        agent_version: String,
-    },
-    /// A task prediction run was begun.
-    TaskBegun {
-        /// Opaque run identifier to pass to compiler shims and commit later.
-        run: String,
-    },
-    /// A task prediction run was committed.
-    TaskCommitted,
-    /// A local CAS path already verified against the requested digest.
-    Blob {
-        /// Verified local path, or `None` on a cache miss.
-        path: Option<PathBuf>,
-    },
-    /// Local CAS paths already verified against the requested digests.
-    Blobs {
-        /// Verified local paths or misses, in request order.
-        paths: Vec<Option<PathBuf>>,
-    },
-    /// A blob was stored locally.
-    Stored {
-        /// Path of the stored object in the local CAS.
-        path: PathBuf,
-    },
-    /// Result of an action lookup.
-    ActionResult {
-        /// Validated action result, or `None` on a cache miss.
-        result: Option<RemoteActionResult>,
-    },
-    /// Hit statistics were updated.
-    ActionHitRecorded,
-    /// Verification statistics were updated.
-    ActionVerificationRecorded,
-    /// Bypass statistics were updated.
-    BypassRecorded,
-    /// Unconsulted-compilation statistics were updated.
-    UnconsultedRecorded,
-    /// Compiler invocation accounting was recorded.
-    CompilerInvocationRecorded,
-    /// An action result was stored.
-    ActionStored {
-        /// Path of the stored local action-result record.
-        path: PathBuf,
-    },
-    /// Result of an input-prediction lookup.
-    ActionPrediction {
-        /// Matching prediction, or `None` when none is known.
-        prediction: Option<ActionPrediction>,
-    },
-    /// An input prediction was recorded.
-    ActionPredictionRecorded,
-    /// Result of an executable-identity lookup.
-    ExecutableIdentity {
-        /// Captured output, or `None` when no identity is cached.
-        stdout: Option<Vec<u8>>,
-    },
-    /// The request failed without terminating the agent connection.
-    Error {
-        /// Human-readable failure description.
-        message: String,
-    },
-    /// A shim diagnostic was accepted for the session to surface.
-    ///
-    /// Sits past `Error` for the reason [`AgentRequest::RecordWarning`] sits
-    /// last: anywhere earlier renumbers the variants below it. New variants go
-    /// here.
-    WarningRecorded,
-    /// Digests recorded earlier for the requested file identities.
-    FileDigests {
-        /// Recorded digests or misses, in request order.
-        digests: Vec<Option<CacheDigest>>,
-    },
-    /// File digests were recorded.
-    FileDigestsRecorded,
-}
-
-/// Aggregate cache activity for one task session.
-///
-/// The agent produces these; nothing outside this crate has cause to build one.
-/// Saying so keeps a new counter from being a breaking change, which is what
-/// this type exists to accumulate -- reach for [`AgentStats::default`] and
-/// assign the fields a test needs.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct AgentStats {
-    /// End-to-end lifetime of the task-scoped cache session.
-    pub session_duration_ns: u64,
-    /// Number of action-result lookups.
-    pub lookups: u64,
-    /// Compilations no action-result lookup was possible for, because no usable
-    /// action key was available.
-    ///
-    /// Counted separately from a miss, which is a lookup that found nothing.
-    /// Both compile, but only a miss says a lookup happened.
-    pub unconsulted: u64,
-    /// Number of lookups that found a valid local action result.
-    pub hits: u64,
-    /// Number of newly stored content-addressed objects.
-    pub stores: u64,
-    /// Total size of newly stored objects.
-    pub stored_bytes: u64,
-    /// Number of cache hits compiled again for qualification.
-    pub verifications: u64,
-    /// Number of qualification builds that diverged from the cached result.
-    pub divergences: u64,
-    /// CAS payload bytes downloaded from the remote cache.
-    pub downloaded_bytes: u64,
-    /// CAS payload bytes uploaded to the remote cache.
-    pub uploaded_bytes: u64,
-    /// Objects published to the remote cache after the build asked for them.
-    ///
-    /// A store request returns once the object is in the local CAS, so the
-    /// upload it implies happens off the build's critical path. This counts the
-    /// blobs and action results that publication actually completed for.
-    pub background_uploads: u64,
-    /// Queued uploads that did not publish, having been reported and recovered
-    /// from.
-    pub background_upload_failures: u64,
-    /// Framed requests that published several blobs at once.
-    pub remote_blob_pack_uploads: u64,
-    /// Blobs published through those framed requests.
-    pub remote_blob_pack_upload_blobs: u64,
-    /// Time the session spent waiting for queued uploads once the build ended.
-    pub upload_drain_duration_ns: u64,
-    /// Complete actions staged before an adapter requested them.
-    pub prefetched_actions: u64,
-    /// Predictions carried by the task manifest this session loaded.
-    ///
-    /// Zero means no earlier build left a manifest behind -- a genuinely cold
-    /// start. Together with `lookups`, this is what tells a session that loaded
-    /// hundreds of predictions and matched none of them apart from one that had
-    /// nothing to match against; the first means the invocations changed (a
-    /// compiler update does this to every one of them at once), the second that
-    /// the store was empty.
-    pub predictions_loaded: u64,
-    /// Compilations that were not cacheable, counted by reason.
-    pub bypasses: BTreeMap<String, u64>,
-    /// Estimated compiler time avoided by restored action hits.
-    pub avoided_compiler_duration_ns: u64,
-    /// Real compiler work performed in this session, grouped by outcome.
-    pub compiler: BTreeMap<String, CompilerStats>,
-    /// Cumulative real compiler time by crate name.
-    pub slow_compilations: BTreeMap<String, u64>,
-    /// Remote cache operations that failed and were degraded to a local result.
-    ///
-    /// A remote cache that cannot be reached, or that answers in a way this
-    /// client refuses, costs hit rate rather than correctness, so every one of
-    /// these is recovered from rather than raised. Counting them is what keeps a
-    /// remote that is failing every request from reading as one that merely had
-    /// nothing to offer.
-    pub remote_failures: u64,
-    /// Number of task manifest requests made to the remote cache.
-    pub remote_manifest_lookups: u64,
-    /// Cumulative time spent requesting remote task manifests.
-    pub remote_manifest_lookup_duration_ns: u64,
-    /// Number of action-result requests made to the remote cache.
-    pub remote_action_lookups: u64,
-    /// Cumulative time spent requesting remote action results.
-    pub remote_action_lookup_duration_ns: u64,
-    /// Number of blob requests made to the remote cache.
-    pub remote_blob_requests: u64,
-    /// Number of packed blob requests made to the remote cache.
-    pub remote_blob_pack_requests: u64,
-    /// Number of verified blobs received through packed responses.
-    pub remote_blob_pack_blobs: u64,
-    /// Cumulative time spent downloading and verifying remote blobs.
-    pub remote_blob_transfer_duration_ns: u64,
-    /// Cumulative time spent ingesting downloaded blobs into the local CAS.
-    pub local_cas_write_duration_ns: u64,
-    /// Number of speculative prefetch runs started for task manifests.
-    pub prefetch_runs: u64,
-    /// Cumulative wall time of speculative task-manifest prefetch runs.
-    pub prefetch_duration_ns: u64,
-    /// Cumulative time spent staging or materializing and validating cached outputs.
-    pub materialization_duration_ns: u64,
-    /// Number of compiler output files restored from action hits.
-    pub restored_output_files: u64,
-    /// Declared size of compiler output files restored from action hits.
-    pub restored_output_bytes: u64,
-    /// Number of restored output files materialized with filesystem reflinks.
-    pub reflinked_output_files: u64,
-    /// Declared size of outputs materialized with filesystem reflinks.
-    pub reflinked_output_bytes: u64,
-    /// Number of restored output files materialized by copying their bytes.
-    pub copied_output_files: u64,
-    /// Declared size of outputs materialized by copying their bytes.
-    pub copied_output_bytes: u64,
-    /// Number of output files kept in place, already holding the cached bytes.
-    pub reused_output_files: u64,
-    /// Declared size of outputs kept in place.
-    pub reused_output_bytes: u64,
-}
-
-/// Count and cumulative wall time for one compiler-invocation outcome.
-///
-/// Non-exhaustive for the same reason as [`AgentStats`], which holds these:
-/// leaving the outer bag open does not help if describing an outcome in more
-/// detail still breaks the type inside it.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct CompilerStats {
-    /// Number of compiler invocations observed.
-    pub invocations: u64,
-    /// Cumulative wall time spent in those invocations.
-    pub duration_ns: u64,
-}
-
-impl CompilerStats {
-    /// The counts observed for one outcome.
-    pub fn new(invocations: u64, duration_ns: u64) -> Self {
-        Self {
-            invocations,
-            duration_ns,
-        }
-    }
 }
 
 #[derive(Default)]
@@ -705,106 +264,6 @@ impl UploadSink for AgentUploadSink {
             .fetch_add(1, Ordering::Relaxed);
         self.stats.remote_failures.fetch_add(1, Ordering::Relaxed);
     }
-}
-
-/// The on-disk identity a recorded file digest describes.
-///
-/// The same trade [`VerifiedBlob`] makes for CAS reads, offered to shims for
-/// the files they hash: an overwrite moves the modification time and a
-/// truncation changes the length, so a digest recorded against both stands
-/// until either does. Where the platform reports a metadata-change time the
-/// identity carries that too, and it is the part a writer cannot restore: a
-/// rewrite that puts the modification time back still moves the change time,
-/// so only filesystems without one fall back to the freshness model the
-/// surrounding build tool already lives on.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct FileIdentity {
-    /// Absolute path of the file.
-    pub path: PathBuf,
-    /// Length of the file in bytes.
-    pub len: u64,
-    /// Modification time of the file.
-    pub modified: SystemTime,
-    /// Platform metadata-change token, where one exists.
-    pub changed: Option<(i64, i64)>,
-}
-
-impl FileIdentity {
-    /// Describe a file from metadata already in hand, or nothing when the
-    /// filesystem reports no modification time to compare against later.
-    pub fn describe(path: &Path, metadata: &std::fs::Metadata) -> Option<Self> {
-        Some(Self {
-            path: path.to_path_buf(),
-            len: metadata.len(),
-            modified: metadata.modified().ok()?,
-            changed: change_token(metadata),
-        })
-    }
-}
-
-/// The metadata-change time as an opaque token, where the platform has one.
-#[cfg(unix)]
-fn change_token(metadata: &std::fs::Metadata) -> Option<(i64, i64)> {
-    use std::os::unix::fs::MetadataExt;
-    Some((metadata.ctime(), metadata.ctime_nsec()))
-}
-
-/// Windows reports creation rather than metadata-change time, which a rewrite
-/// does not move, so no token is better than a misleading one.
-#[cfg(not(unix))]
-fn change_token(_metadata: &std::fs::Metadata) -> Option<(i64, i64)> {
-    None
-}
-
-/// A file digest recorded against the identity it was read under.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct RecordedFileDigest {
-    /// Identity of the file when its contents were hashed.
-    pub file: FileIdentity,
-    /// Digest of those contents.
-    pub digest: CacheDigest,
-}
-
-/// What a recorded file digest may stand in for.
-///
-/// Adapters prove different things when they read a file: the cc adapter's
-/// input scan also establishes that a source embeds no timestamp macro, which
-/// a digest recorded by the rustc adapter never checked. Scoping the ledger
-/// keeps one adapter's shortcut from resting on a property another adapter
-/// never established.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum FileDigestScope {
-    /// The digest describes the file's contents and nothing more.
-    Content,
-    /// The digest describes a cc compiler input that also passed the
-    /// timestamp-macro scan.
-    CcInput,
-}
-
-/// Recorded digests a session may consult instead of rehashing a file.
-///
-/// The agent's file-digest ledger answers through this everywhere a caller
-/// holds one; the no-op implementation stands in where reuse must not happen,
-/// such as under verification.
-pub trait FileDigestCache: Send + Sync {
-    /// Recorded digests for these identities, in request order.
-    fn find(&self, scope: FileDigestScope, files: &[FileIdentity]) -> Vec<Option<CacheDigest>>;
-    /// Record digests for files read under these identities.
-    fn record(&self, scope: FileDigestScope, entries: Vec<RecordedFileDigest>);
-}
-
-/// A [`FileDigestCache`] that remembers nothing and finds nothing.
-pub struct NoFileDigestCache;
-
-impl FileDigestCache for NoFileDigestCache {
-    fn find(&self, _scope: FileDigestScope, files: &[FileIdentity]) -> Vec<Option<CacheDigest>> {
-        vec![None; files.len()]
-    }
-
-    fn record(&self, _scope: FileDigestScope, _entries: Vec<RecordedFileDigest>) {}
 }
 
 /// A CAS blob whose contents this session hashed, and the file identity that
@@ -1513,810 +972,6 @@ impl CacheAgent {
         lock
     }
 
-    fn spawn_prefetch_predictions(&self, predictions: Vec<ActionPrediction>) {
-        if predictions.is_empty() || !self.remote_mode.reads() || self.remote.is_none() {
-            return;
-        }
-        let agent = self.clone();
-        let task = tokio::spawn(async move {
-            agent.prefetch_predictions(predictions.iter()).await;
-        });
-        self.prefetch_tasks.lock().unwrap().push(task);
-    }
-
-    async fn prefetch_predictions<'a>(
-        &self,
-        predictions: impl Iterator<Item = &'a ActionPrediction>,
-    ) {
-        if !self.remote_mode.reads() || self.remote.is_none() {
-            return;
-        }
-        self.stats.prefetch_runs.fetch_add(1, Ordering::Relaxed);
-        let _timer = AtomicDurationTimer::start(&self.stats.prefetch_duration_ns);
-        let mut actions = BTreeMap::new();
-        for prediction in predictions {
-            actions
-                .entry(prediction.action.clone())
-                .or_insert_with(|| prediction.adapter.clone());
-        }
-        // One request per predicted action is the bulk of a prefetch's latency on
-        // a large workspace, so ask for them together where the server allows it.
-        match self.prefetch_action_batches(&actions).await {
-            Ok(true) => return,
-            Ok(false) => {}
-            Err(error) => {
-                self.note_remote_failure();
-                warn!("remote action batch lookup failed: {error}");
-            }
-        }
-        let mut actions = actions.into_iter();
-        let mut tasks = tokio::task::JoinSet::new();
-        for _ in 0..MAX_PREFETCH_TRANSFERS {
-            let Some((action, adapter)) = actions.next() else {
-                break;
-            };
-            let agent = self.clone();
-            tasks.spawn(async move { agent.resolve_prefetch_action(action, adapter).await });
-        }
-        let mut resolved = Vec::new();
-        while !tasks.is_empty() {
-            let result = if resolved.is_empty() {
-                tasks.join_next().await
-            } else {
-                match tokio::time::timeout(PREFETCH_ACTION_BATCH_DELAY, tasks.join_next()).await {
-                    Ok(result) => result,
-                    Err(_) => {
-                        self.prefetch_resolved_actions(std::mem::take(&mut resolved))
-                            .await;
-                        continue;
-                    }
-                }
-            };
-            let Some(result) = result else {
-                break;
-            };
-            match result {
-                Ok(Ok(Some(action))) => resolved.push(action),
-                Ok(Ok(None)) => {}
-                Ok(Err(error)) => {
-                    self.note_remote_failure();
-                    warn!("remote action prefetch failed: {error}");
-                }
-                Err(error) => {
-                    self.note_remote_failure();
-                    warn!("remote action prefetch task failed: {error}");
-                }
-            }
-            if let Some((action, adapter)) = actions.next() {
-                let agent = self.clone();
-                tasks.spawn(async move { agent.resolve_prefetch_action(action, adapter).await });
-            }
-            if resolved.len() == MAX_PREFETCH_ACTION_BATCH {
-                self.prefetch_resolved_actions(std::mem::take(&mut resolved))
-                    .await;
-            }
-        }
-        if !resolved.is_empty() {
-            self.prefetch_resolved_actions(resolved).await;
-        }
-    }
-
-    /// Resolve predicted actions in batched lookups, staging what comes back.
-    ///
-    /// Returns whether the batch extension answered. A server without it, or one
-    /// that stops answering part way through, leaves the per-action path to
-    /// resolve whatever is left -- the results already staged here are memoized,
-    /// so nothing is looked up twice.
-    async fn prefetch_action_batches(
-        &self,
-        actions: &BTreeMap<CacheDigest, String>,
-    ) -> Result<bool> {
-        let Some(remote) = self.remote.as_deref() else {
-            return Ok(false);
-        };
-        let Some(limit) = remote.action_batch_limit().await? else {
-            return Ok(false);
-        };
-        let wanted: Vec<CacheDigest> = actions
-            .keys()
-            .filter(|action| !self.action_is_staged(action))
-            .cloned()
-            .collect();
-        if wanted.is_empty() {
-            return Ok(true);
-        }
-        let mut chunks: Vec<Vec<CacheDigest>> = Vec::new();
-        for chunk in wanted.chunks(limit) {
-            chunks.push(chunk.to_vec());
-        }
-        let mut lookups = stream::iter(chunks)
-            .map(|chunk| async move {
-                let _prefetch_permit = self.prefetch_transfers.acquire().await?;
-                let _permit = self.remote_transfers.acquire().await?;
-                self.stats
-                    .remote_action_lookups
-                    .fetch_add(1, Ordering::Relaxed);
-                let _timer =
-                    AtomicDurationTimer::start(&self.stats.remote_action_lookup_duration_ns);
-                remote.get_action_results(&chunk).await
-            })
-            .buffer_unordered(MAX_PREFETCH_BATCH_LOOKUPS);
-        let mut answered = true;
-        while let Some(lookup) = lookups.next().await {
-            let results = match lookup {
-                Ok(Some(results)) => results,
-                // Advertised but unavailable. What is left falls back.
-                Ok(None) => {
-                    answered = false;
-                    continue;
-                }
-                Err(error) => {
-                    self.note_remote_failure();
-                    warn!("remote action batch lookup failed: {error}");
-                    answered = false;
-                    continue;
-                }
-            };
-            let mut resolved = Vec::with_capacity(results.len());
-            for result in results {
-                let Some(adapter) = actions.get(&result.action).cloned() else {
-                    continue;
-                };
-                let lock = self.action_lock(&result.action);
-                let _guard = lock.lock().await;
-                if self.actions.find(&result.action)?.is_some() {
-                    continue;
-                }
-                self.pending_remote_actions
-                    .lock()
-                    .unwrap()
-                    .insert(result.action.clone(), result.clone());
-                resolved.push(PrefetchedAction { adapter, result });
-            }
-            while !resolved.is_empty() {
-                let wave = resolved
-                    .drain(..resolved.len().min(MAX_PREFETCH_ACTION_BATCH))
-                    .collect();
-                self.prefetch_resolved_actions(wave).await;
-            }
-        }
-        Ok(answered)
-    }
-
-    /// Whether an action's result is already local or already looked up.
-    fn action_is_staged(&self, action: &CacheDigest) -> bool {
-        if self
-            .pending_remote_actions
-            .lock()
-            .unwrap()
-            .contains_key(action)
-        {
-            return true;
-        }
-        self.actions.find(action).is_ok_and(|found| found.is_some())
-    }
-
-    #[cfg(test)]
-    async fn prefetch_action(&self, action: CacheDigest, adapter: String) -> Result<()> {
-        if let Some(action) = self.resolve_prefetch_action(action, adapter).await? {
-            self.prefetch_resolved_actions(vec![action]).await;
-        }
-        Ok(())
-    }
-
-    async fn resolve_prefetch_action(
-        &self,
-        action: CacheDigest,
-        adapter: String,
-    ) -> Result<Option<PrefetchedAction>> {
-        let remote = self
-            .remote
-            .as_ref()
-            .ok_or_else(|| eyre::eyre!("remote cache is not configured"))?;
-        let result = {
-            let lock = self.action_lock(&action);
-            let _guard = lock.lock().await;
-            if self.actions.find(&action)?.is_some() {
-                return Ok(None);
-            }
-            if let Some(result) = self
-                .pending_remote_actions
-                .lock()
-                .unwrap()
-                .get(&action)
-                .cloned()
-            {
-                result
-            } else {
-                let _prefetch_permit = self.prefetch_transfers.acquire().await?;
-                let result = {
-                    let _permit = self.remote_transfers.acquire().await?;
-                    self.get_remote_action_result(remote, &action).await?
-                };
-                let Some(result) = result else {
-                    return Ok(None);
-                };
-                self.pending_remote_actions
-                    .lock()
-                    .unwrap()
-                    .insert(action.clone(), result.clone());
-                result
-            }
-        };
-        Ok(Some(PrefetchedAction { adapter, result }))
-    }
-
-    fn prefetch_resolved_actions(&self, actions: Vec<PrefetchedAction>) -> BoxFuture<'_, ()> {
-        self.prefetch_resolved_actions_inner(actions).boxed()
-    }
-
-    async fn prefetch_resolved_actions_inner(&self, actions: Vec<PrefetchedAction>) {
-        let Some(remote) = self.remote.as_deref() else {
-            return;
-        };
-        if actions.is_empty() {
-            return;
-        }
-
-        let mut top_level = BTreeMap::new();
-        for action in &actions {
-            for digest in [
-                Some(&action.result.action),
-                action.result.metadata.as_ref(),
-                action.result.output_root.as_ref(),
-            ]
-            .into_iter()
-            .flatten()
-            {
-                top_level.insert(digest.clone(), ());
-            }
-        }
-        let mut verified = self
-            .fetch_remote_blobs(
-                remote,
-                top_level.into_keys().collect(),
-                Some(&self.prefetch_transfers),
-            )
-            .await;
-
-        let mut next = BTreeMap::new();
-        let mut pending_directories = BTreeMap::new();
-        let mut parsed_directories = BTreeMap::new();
-        let mut rustc_metadata = BTreeMap::new();
-        for action in &actions {
-            if action.adapter == "rustc"
-                && let Some(metadata_digest) = &action.result.metadata
-            {
-                match verified
-                    .get(metadata_digest)
-                    .ok_or_else(|| eyre::eyre!("remote rustc action metadata is missing"))
-                    .and_then(|path| Self::parse_rustc_metadata(path))
-                {
-                    Ok(metadata) => {
-                        queue_prefetch_digest(&verified, &mut next, metadata.stdout.clone());
-                        queue_prefetch_digest(&verified, &mut next, metadata.stderr.clone());
-                        rustc_metadata.insert(metadata_digest.clone(), metadata);
-                    }
-                    Err(error) => warn!(
-                        "remote rustc action metadata prefetch failed for {}: {error}",
-                        action.result.action.hash
-                    ),
-                }
-            }
-            if let Some(output_root) = &action.result.output_root {
-                pending_directories.insert(output_root.clone(), ());
-            }
-        }
-
-        let mut seen_directories = BTreeMap::new();
-        loop {
-            let mut following = BTreeMap::new();
-            let mut directory_limit_exceeded = false;
-            for digest in pending_directories.into_keys() {
-                following.remove(&digest);
-                if seen_directories.insert(digest.clone(), ()).is_some() {
-                    continue;
-                }
-                if seen_directories.len() > MAX_PREFETCH_DIRECTORY_OBJECTS {
-                    warn!("remote action output tree is too large to prefetch");
-                    following.clear();
-                    break;
-                }
-                match verified
-                    .get(&digest)
-                    .ok_or_else(|| eyre::eyre!("remote action output directory is missing"))
-                    .and_then(|path| Self::parse_cache_directory(path))
-                {
-                    Ok(directory) => {
-                        for file in &directory.files {
-                            queue_prefetch_digest(&verified, &mut next, file.digest.clone());
-                            if next.len() >= MAX_PREFETCH_OBJECTS_PER_WAVE {
-                                self.flush_prefetch_digest_batch(remote, &mut verified, &mut next)
-                                    .await;
-                            }
-                        }
-                        for child in &directory.directories {
-                            if !queue_prefetch_directory(
-                                &seen_directories,
-                                &mut following,
-                                child.digest.clone(),
-                                MAX_PREFETCH_DIRECTORY_OBJECTS,
-                            ) {
-                                warn!("remote action output tree is too large to prefetch");
-                                directory_limit_exceeded = true;
-                                break;
-                            }
-                            queue_prefetch_digest(&verified, &mut next, child.digest.clone());
-                            if next.len() >= MAX_PREFETCH_OBJECTS_PER_WAVE {
-                                self.flush_prefetch_digest_batch(remote, &mut verified, &mut next)
-                                    .await;
-                            }
-                        }
-                        parsed_directories.insert(digest, directory);
-                    }
-                    Err(error) => warn!(
-                        "remote action output directory prefetch failed for {}: {error}",
-                        digest.hash
-                    ),
-                }
-                if directory_limit_exceeded {
-                    following.clear();
-                    break;
-                }
-            }
-            self.flush_prefetch_digest_batch(remote, &mut verified, &mut next)
-                .await;
-            if following.is_empty() {
-                break;
-            }
-            pending_directories = following;
-        }
-
-        for action in actions {
-            match Self::validate_prefetched_action(
-                &action,
-                &verified,
-                &rustc_metadata,
-                &parsed_directories,
-            ) {
-                Ok(()) => {
-                    if let Err(error) = self.actions.store(&action.result) {
-                        warn!(
-                            "remote action prefetch could not publish {}: {error}",
-                            action.result.action.hash
-                        );
-                        continue;
-                    }
-                    self.pending_remote_actions
-                        .lock()
-                        .unwrap()
-                        .remove(&action.result.action);
-                    self.stats
-                        .prefetched_actions
-                        .fetch_add(1, Ordering::Relaxed);
-                }
-                Err(error) => warn!(
-                    "remote action prefetch was incomplete for {}: {error}",
-                    action.result.action.hash
-                ),
-            }
-        }
-    }
-
-    async fn flush_prefetch_digest_batch(
-        &self,
-        remote: &RemoteCacheClient,
-        verified: &mut BTreeMap<CacheDigest, PathBuf>,
-        pending: &mut BTreeMap<CacheDigest, ()>,
-    ) {
-        if pending.is_empty() {
-            return;
-        }
-        let digests = std::mem::take(pending).into_keys().collect();
-        verified.extend(
-            self.fetch_remote_blobs(remote, digests, Some(&self.prefetch_transfers))
-                .await,
-        );
-    }
-
-    async fn fetch_remote_blobs(
-        &self,
-        remote: &RemoteCacheClient,
-        digests: Vec<CacheDigest>,
-        prefetch_limit: Option<&tokio::sync::Semaphore>,
-    ) -> BTreeMap<CacheDigest, PathBuf> {
-        let mut verified = BTreeMap::new();
-        let mut missing = BTreeMap::new();
-        for digest in digests {
-            match self.find_verified_blob(&digest) {
-                Ok(Some(path)) => {
-                    verified.insert(digest, path);
-                }
-                Ok(None) => {
-                    missing.insert(digest, ());
-                }
-                Err(error) => warn!(
-                    "local cache blob lookup failed for {}: {error}",
-                    digest.hash
-                ),
-            }
-        }
-        if missing.is_empty() {
-            return verified;
-        }
-
-        let mut pack_candidates = missing.clone();
-        while !pack_candidates.is_empty() {
-            let candidates = match blob_pack_chunk(
-                &pack_candidates.keys().cloned().collect::<Vec<_>>(),
-                BlobPackLimits {
-                    max_items: MAX_STAGED_BLOB_PACK_ITEMS,
-                    max_bytes: MAX_STAGED_BLOB_PACK_BYTES,
-                },
-            ) {
-                Ok(candidates) if !candidates.is_empty() => candidates,
-                Ok(_) => break,
-                Err(error) => {
-                    warn!("remote cache blob pack skipped: {error}");
-                    break;
-                }
-            };
-            // A pack and an individual fetch share these per-digest locks. Hold
-            // them through ingestion so overlapping prefetch and foreground
-            // requests cannot download or charge the same object twice.
-            let mut pack_guards = BTreeMap::new();
-            for digest in candidates {
-                let guard = self.write_lock(&digest).lock_owned().await;
-                match self.find_verified_blob(&digest) {
-                    Ok(Some(path)) => {
-                        pack_candidates.remove(&digest);
-                        missing.remove(&digest);
-                        verified.insert(digest, path);
-                    }
-                    Ok(None) => {
-                        pack_guards.insert(digest, guard);
-                    }
-                    Err(error) => {
-                        warn!(
-                            "local cache blob lookup failed for {}: {error}",
-                            digest.hash
-                        );
-                        pack_guards.insert(digest, guard);
-                    }
-                }
-            }
-            let requested = pack_guards.keys().cloned().collect::<Vec<_>>();
-            if requested.is_empty() {
-                continue;
-            }
-            let requested_bytes = requested
-                .iter()
-                .fold(0_u64, |total, digest| total.saturating_add(digest.size));
-            let pack_reservation = match self
-                .reserve_remote_download_up_to(requested_bytes.min(MAX_STAGED_BLOB_PACK_BYTES))
-            {
-                Ok(reservation) if reservation.bytes() > 0 => reservation,
-                Ok(_) => break,
-                Err(error) => {
-                    warn!("remote cache blob pack skipped: {error}");
-                    break;
-                }
-            };
-            let (pack, transfer_duration_ns) = {
-                let _prefetch_permit = match prefetch_limit {
-                    Some(limit) => match limit.acquire().await {
-                        Ok(permit) => Some(permit),
-                        Err(error) => {
-                            warn!(
-                                "remote cache blob pack could not acquire prefetch limit: {error}"
-                            );
-                            break;
-                        }
-                    },
-                    None => None,
-                };
-                let _transfer_permit = match self.remote_transfers.acquire().await {
-                    Ok(permit) => permit,
-                    Err(error) => {
-                        warn!("remote cache blob pack could not acquire transfer limit: {error}");
-                        break;
-                    }
-                };
-                let transfer_started = Instant::now();
-                let pack = remote
-                    .get_blob_pack_with_limit(
-                        &requested,
-                        self.remote_staging_dir.as_path(),
-                        pack_reservation.bytes(),
-                    )
-                    .await;
-                (pack, duration_ns(transfer_started))
-            };
-            let pack = match pack {
-                Ok(Some(pack)) => pack,
-                Ok(None) => break,
-                Err(error) => {
-                    atomic_saturating_add(
-                        &self.stats.remote_blob_transfer_duration_ns,
-                        transfer_duration_ns,
-                    );
-                    warn!(
-                        "remote cache blob pack failed; falling back to individual blobs: {error}"
-                    );
-                    break;
-                }
-            };
-            pack_reservation.commit(pack.payload_bytes);
-            atomic_saturating_add(
-                &self.stats.remote_blob_transfer_duration_ns,
-                transfer_duration_ns,
-            );
-            atomic_saturating_add(&self.stats.remote_blob_pack_requests, pack.requests);
-            atomic_saturating_add(&self.stats.remote_blob_pack_blobs, pack.blob_count);
-            atomic_saturating_add(&self.stats.downloaded_bytes, pack.payload_bytes);
-            if pack.requested.is_empty() {
-                // The server's negotiated cap can be smaller than the local
-                // staging cap used to select this locked slice. Fall back for
-                // this slice, but keep packing later candidates that may fit.
-                for digest in &requested {
-                    pack_candidates.remove(digest);
-                }
-                continue;
-            }
-            for digest in &pack.requested {
-                pack_candidates.remove(digest);
-            }
-            let mut ingests = stream::iter(pack.blobs.into_iter().map(|(digest, source)| {
-                let digest_for_result = digest.clone();
-                let guard = pack_guards
-                    .remove(&digest)
-                    .expect("requested packed blob has a write lock");
-                async move {
-                    (
-                        digest_for_result,
-                        self.ingest_packed_blob(digest, source, guard).await,
-                    )
-                }
-            }))
-            .buffer_unordered(MAX_PREFETCH_TRANSFERS);
-            while let Some((digest, result)) = ingests.next().await {
-                match result {
-                    Ok(path) => {
-                        missing.remove(&digest);
-                        verified.insert(digest, path);
-                    }
-                    Err(error) => warn!(
-                        "remote cache packed blob ingest failed for {}: {error}",
-                        digest.hash
-                    ),
-                }
-            }
-        }
-
-        let mut transfers = stream::iter(missing.into_keys().map(|digest| {
-            let digest_for_result = digest.clone();
-            async move {
-                (
-                    digest_for_result,
-                    self.fetch_remote_blob_with_limit(remote, &digest, prefetch_limit)
-                        .await,
-                )
-            }
-        }))
-        .buffer_unordered(MAX_PREFETCH_TRANSFERS);
-        while let Some((digest, result)) = transfers.next().await {
-            match result {
-                Ok(path) => {
-                    verified.insert(digest, path);
-                }
-                Err(error) => warn!(
-                    "remote cache blob prefetch failed for {}: {error}",
-                    digest.hash
-                ),
-            }
-        }
-        verified
-    }
-
-    async fn ingest_packed_blob(
-        &self,
-        digest: CacheDigest,
-        source: PathBuf,
-        _guard: tokio::sync::OwnedMutexGuard<()>,
-    ) -> Result<PathBuf> {
-        let digest_size = digest.size;
-        let agent = self.clone();
-        let (path, stored, cas_duration_ns) = tokio::task::spawn_blocking(move || {
-            if let Some(path) = agent.find_verified_blob(&digest)? {
-                return Ok::<_, eyre::Report>((path, false, 0));
-            }
-            let cas_started = Instant::now();
-            let path = agent.cas.store_verified_file(&digest, &source)?;
-            let cas_duration_ns = duration_ns(cas_started);
-            agent.remember_verified_blob(&digest, &path);
-            Ok((path, true, cas_duration_ns))
-        })
-        .await??;
-        atomic_saturating_add(&self.stats.local_cas_write_duration_ns, cas_duration_ns);
-        if stored {
-            self.stats.stores.fetch_add(1, Ordering::Relaxed);
-            atomic_saturating_add(&self.stats.stored_bytes, digest_size);
-        }
-        Ok(path)
-    }
-
-    fn parse_rustc_metadata(path: &Path) -> Result<RustcMetadata> {
-        let bytes = fs::read(path)?;
-        let metadata: RustcMetadata = serde_json::from_slice(&bytes)?;
-        if metadata.version != 1 || metadata.kind != "rustc" || canonical_json(&metadata)? != bytes
-        {
-            bail!("remote rustc action metadata is invalid");
-        }
-        Ok(metadata)
-    }
-
-    fn parse_cache_directory(path: &Path) -> Result<CacheDirectory> {
-        let bytes = fs::read(path)?;
-        let directory: CacheDirectory = serde_json::from_slice(&bytes)?;
-        if directory.version != 1 || canonical_json(&directory)? != bytes {
-            bail!("remote action output directory is invalid");
-        }
-        Ok(directory)
-    }
-
-    #[cfg(test)]
-    fn load_cache_directory(&self, digest: &CacheDigest) -> Result<CacheDirectory> {
-        let path = self
-            .find_verified_blob(digest)?
-            .ok_or_else(|| eyre::eyre!("remote action output directory is missing"))?;
-        Self::parse_cache_directory(&path)
-    }
-
-    fn validate_prefetched_action(
-        action: &PrefetchedAction,
-        verified: &BTreeMap<CacheDigest, PathBuf>,
-        rustc_metadata: &BTreeMap<CacheDigest, RustcMetadata>,
-        directories: &BTreeMap<CacheDigest, CacheDirectory>,
-    ) -> Result<()> {
-        if !verified.contains_key(&action.result.action) {
-            bail!("remote action descriptor is missing");
-        }
-        if let Some(metadata) = &action.result.metadata {
-            if action.adapter == "rustc" {
-                let metadata = rustc_metadata
-                    .get(metadata)
-                    .ok_or_else(|| eyre::eyre!("remote rustc action metadata is missing"))?;
-                for digest in [&metadata.stdout, &metadata.stderr] {
-                    if !verified.contains_key(digest) {
-                        bail!("remote rustc action diagnostic blob is missing");
-                    }
-                }
-            } else if !verified.contains_key(metadata) {
-                bail!("remote action metadata is missing");
-            }
-        }
-        let mut pending = action
-            .result
-            .output_root
-            .iter()
-            .cloned()
-            .collect::<Vec<_>>();
-        let mut seen = BTreeMap::new();
-        while let Some(digest) = pending.pop() {
-            if seen.insert(digest.clone(), ()).is_some() {
-                continue;
-            }
-            if seen.len() > MAX_PREFETCH_DIRECTORY_OBJECTS {
-                bail!("remote action output tree is too large");
-            }
-            let directory = directories
-                .get(&digest)
-                .ok_or_else(|| eyre::eyre!("remote action output directory is missing"))?;
-            for file in &directory.files {
-                if !verified.contains_key(&file.digest) {
-                    bail!("remote action output file is missing");
-                }
-            }
-            pending.extend(
-                directory
-                    .directories
-                    .iter()
-                    .map(|directory| directory.digest.clone()),
-            );
-        }
-        Ok(())
-    }
-
-    #[cfg(test)]
-    async fn prefetch_output_tree(
-        &self,
-        remote: &RemoteCacheClient,
-        output_root: &CacheDigest,
-    ) -> Result<()> {
-        let mut pending = vec![output_root.clone()];
-        let mut seen = BTreeMap::new();
-        while let Some(digest) = pending.pop() {
-            if seen.insert(digest.clone(), ()).is_some() {
-                continue;
-            }
-            if seen.len() > MAX_PREFETCH_DIRECTORY_OBJECTS {
-                bail!("remote action output tree is too large");
-            }
-            self.fetch_remote_blob_with_limit(remote, &digest, Some(&self.prefetch_transfers))
-                .await?;
-            let directory = self.load_cache_directory(&digest)?;
-            let mut transfers = stream::iter(directory.files.into_iter().map(|file| async move {
-                self.fetch_remote_blob_with_limit(
-                    remote,
-                    &file.digest,
-                    Some(&self.prefetch_transfers),
-                )
-                .await
-                .map(|_| ())
-            }))
-            .buffer_unordered(MAX_PREFETCH_TRANSFERS);
-            while let Some(result) = transfers.next().await {
-                result?;
-            }
-            pending.extend(
-                directory
-                    .directories
-                    .into_iter()
-                    .map(|directory| directory.digest),
-            );
-        }
-        Ok(())
-    }
-
-    async fn fetch_remote_blob(
-        &self,
-        remote: &RemoteCacheClient,
-        digest: &CacheDigest,
-    ) -> Result<PathBuf> {
-        self.fetch_remote_blob_with_limit(remote, digest, None)
-            .await
-    }
-
-    async fn fetch_remote_blob_with_limit(
-        &self,
-        remote: &RemoteCacheClient,
-        digest: &CacheDigest,
-        prefetch_limit: Option<&tokio::sync::Semaphore>,
-    ) -> Result<PathBuf> {
-        let lock = self.write_lock(digest);
-        let _guard = lock.lock().await;
-        if let Some(path) = self.find_verified_blob(digest)? {
-            return Ok(path);
-        }
-        let _prefetch_permit = match prefetch_limit {
-            Some(limit) => Some(limit.acquire().await?),
-            None => None,
-        };
-        let _permit = self.remote_transfers.acquire().await?;
-        let reservation = self.reserve_remote_download(digest.size)?;
-        self.stats
-            .remote_blob_requests
-            .fetch_add(1, Ordering::Relaxed);
-        let transfer_timer =
-            AtomicDurationTimer::start(&self.stats.remote_blob_transfer_duration_ns);
-        let temporary = remote
-            .get_blob_file(digest, self.remote_staging_dir.as_path())
-            .await?;
-        drop(transfer_timer);
-        let _cas_timer = AtomicDurationTimer::start(&self.stats.local_cas_write_duration_ns);
-        let path = self.cas.store_verified_file(digest, temporary.path())?;
-        reservation.commit(digest.size);
-        self.remember_verified_blob(digest, &path);
-        self.stats.stores.fetch_add(1, Ordering::Relaxed);
-        self.stats
-            .stored_bytes
-            .fetch_add(digest.size, Ordering::Relaxed);
-        self.stats
-            .downloaded_bytes
-            .fetch_add(digest.size, Ordering::Relaxed);
-        Ok(path)
-    }
-
     /// Answer one request outside any connection.
     ///
     /// An upload queued this way has no sibling requests to order against, so
@@ -2991,112 +1646,6 @@ where
         Ok(None)
     } else {
         Ok(Some(String::from_utf8(line)?))
-    }
-}
-
-/// Whether `task` is a well-formed task action identity.
-///
-/// Identities name files and directories in the store, so anything that reads
-/// the store back has to be able to tell an identity from whatever else a user
-/// left lying there.
-pub fn is_task_identity(task: &str) -> bool {
-    task.len() == 64
-        && task
-            .bytes()
-            .all(|byte| byte.is_ascii_digit() || (b'a'..=b'f').contains(&byte))
-}
-
-fn validate_task_identity(task: &str) -> Result<()> {
-    if !is_task_identity(task) {
-        bail!("invalid task action identity");
-    }
-    Ok(())
-}
-
-/// Where a store keeps its task prediction manifests.
-fn task_manifest_dir(store: &Path) -> PathBuf {
-    store.join("task-manifests").join("v1")
-}
-
-/// The action digests a task's prediction manifest recorded.
-///
-/// Read straight off disk rather than through an agent, because a collector
-/// needs the action set of tasks no session is running. A manifest that is
-/// missing or no longer parseable yields no actions rather than an error: this
-/// is a prediction index, so the worst a thin answer costs is a cold prefetch,
-/// or an object collected earlier than it deserved.
-pub fn task_manifest_actions(store: &Path, task: &str) -> Result<Vec<CacheDigest>> {
-    validate_task_identity(task)?;
-    let path = task_manifest_dir(store).join(format!("{task}.json"));
-    let bytes = match fs::read(&path) {
-        Ok(bytes) => bytes,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(Vec::new()),
-        Err(error) => {
-            return Err(error).wrap_err_with(|| format!("failed to read {}", path.display()));
-        }
-    };
-    let Ok(manifest) = serde_json::from_slice::<TaskActionManifest>(&bytes) else {
-        return Ok(Vec::new());
-    };
-    if validate_task_manifest(&manifest, task).is_err() {
-        return Ok(Vec::new());
-    }
-    Ok(manifest
-        .predictions
-        .into_iter()
-        .map(|prediction| prediction.action)
-        .collect())
-}
-
-fn validate_task_manifest(manifest: &TaskActionManifest, task: &str) -> Result<()> {
-    if manifest.task == task && manifest.validate() {
-        Ok(())
-    } else {
-        bail!("invalid task action manifest")
-    }
-}
-
-fn merge_task_manifests(
-    task: &str,
-    base: Option<TaskActionManifest>,
-    update: TaskActionManifest,
-) -> Result<TaskActionManifest> {
-    validate_task_manifest(&update, task)?;
-    let mut predictions = BTreeMap::new();
-    if let Some(base) = base {
-        validate_task_manifest(&base, task)?;
-        predictions.extend(
-            base.predictions
-                .into_iter()
-                .map(|prediction| (prediction.invocation.clone(), prediction)),
-        );
-    }
-    predictions.extend(
-        update
-            .predictions
-            .into_iter()
-            .map(|prediction| (prediction.invocation.clone(), prediction)),
-    );
-    let manifest = TaskActionManifest {
-        version: TASK_ACTION_MANIFEST_VERSION,
-        task: task.to_owned(),
-        predictions: predictions.into_values().collect(),
-    };
-    validate_task_manifest(&manifest, task)?;
-    Ok(manifest)
-}
-
-fn merge_remote_task_manifest(
-    task: &str,
-    remote: TaskActionManifest,
-    local: TaskActionManifest,
-) -> (TaskActionManifest, bool) {
-    match merge_task_manifests(task, Some(remote), local.clone()) {
-        Ok(manifest) => (manifest, true),
-        Err(error) => {
-            warn!("remote task action manifest merge failed for {task}: {error}");
-            (local, false)
-        }
     }
 }
