+++ crates/mbx-cache-core/src/agent/file_digest.rs
@@ -1,0 +1,104 @@
+use crate::CacheDigest;
+use serde::{Deserialize, Serialize};
+use std::path::{Path, PathBuf};
+use std::time::SystemTime;
+
+/// The on-disk identity a recorded file digest describes.
+///
+/// The same trade [`VerifiedBlob`] makes for CAS reads, offered to shims for
+/// the files they hash: an overwrite moves the modification time and a
+/// truncation changes the length, so a digest recorded against both stands
+/// until either does. Where the platform reports a metadata-change time the
+/// identity carries that too, and it is the part a writer cannot restore: a
+/// rewrite that puts the modification time back still moves the change time,
+/// so only filesystems without one fall back to the freshness model the
+/// surrounding build tool already lives on.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct FileIdentity {
+    /// Absolute path of the file.
+    pub path: PathBuf,
+    /// Length of the file in bytes.
+    pub len: u64,
+    /// Modification time of the file.
+    pub modified: SystemTime,
+    /// Platform metadata-change token, where one exists.
+    pub changed: Option<(i64, i64)>,
+}
+
+impl FileIdentity {
+    /// Describe a file from metadata already in hand, or nothing when the
+    /// filesystem reports no modification time to compare against later.
+    pub fn describe(path: &Path, metadata: &std::fs::Metadata) -> Option<Self> {
+        Some(Self {
+            path: path.to_path_buf(),
+            len: metadata.len(),
+            modified: metadata.modified().ok()?,
+            changed: change_token(metadata),
+        })
+    }
+}
+
+/// The metadata-change time as an opaque token, where the platform has one.
+#[cfg(unix)]
+fn change_token(metadata: &std::fs::Metadata) -> Option<(i64, i64)> {
+    use std::os::unix::fs::MetadataExt;
+    Some((metadata.ctime(), metadata.ctime_nsec()))
+}
+
+/// Windows reports creation rather than metadata-change time, which a rewrite
+/// does not move, so no token is better than a misleading one.
+#[cfg(not(unix))]
+fn change_token(_metadata: &std::fs::Metadata) -> Option<(i64, i64)> {
+    None
+}
+
+/// A file digest recorded against the identity it was read under.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct RecordedFileDigest {
+    /// Identity of the file when its contents were hashed.
+    pub file: FileIdentity,
+    /// Digest of those contents.
+    pub digest: CacheDigest,
+}
+
+/// What a recorded file digest may stand in for.
+///
+/// Adapters prove different things when they read a file: the cc adapter's
+/// input scan also establishes that a source embeds no timestamp macro, which
+/// a digest recorded by the rustc adapter never checked. Scoping the ledger
+/// keeps one adapter's shortcut from resting on a property another adapter
+/// never established.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum FileDigestScope {
+    /// The digest describes the file's contents and nothing more.
+    Content,
+    /// The digest describes a cc compiler input that also passed the
+    /// timestamp-macro scan.
+    CcInput,
+}
+
+/// Recorded digests a session may consult instead of rehashing a file.
+///
+/// The agent's file-digest ledger answers through this everywhere a caller
+/// holds one; the no-op implementation stands in where reuse must not happen,
+/// such as under verification.
+pub trait FileDigestCache: Send + Sync {
+    /// Recorded digests for these identities, in request order.
+    fn find(&self, scope: FileDigestScope, files: &[FileIdentity]) -> Vec<Option<CacheDigest>>;
+    /// Record digests for files read under these identities.
+    fn record(&self, scope: FileDigestScope, entries: Vec<RecordedFileDigest>);
+}
+
+/// A [`FileDigestCache`] that remembers nothing and finds nothing.
+pub struct NoFileDigestCache;
+
+impl FileDigestCache for NoFileDigestCache {
+    fn find(&self, _scope: FileDigestScope, files: &[FileIdentity]) -> Vec<Option<CacheDigest>> {
+        vec![None; files.len()]
+    }
+
+    fn record(&self, _scope: FileDigestScope, _entries: Vec<RecordedFileDigest>) {}
+}
